% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{agg_vec}
\alias{agg_vec}
\title{Create an aggregation vector}
\usage{
agg_vec(x = character(), aggregated = logical(vec_size(x)))
}
\arguments{
\item{x}{The vector of values.}

\item{aggregated}{A logical vector to identify which values are \verb{<aggregated>}.}
}
\description{
\lifecycle{maturing}
}
\details{
An aggregation vector extends usual vectors by adding <aggregated> values.
These vectors are typically produced via the \code{\link[=aggregate_key]{aggregate_key()}} function,
however it can be useful to create them manually to produce more complicated
hierarchies (such as unbalanced hierarchies).
}
\examples{
agg_vec(
  x = c(NA, "A", "B"),
  aggregated = c(TRUE, FALSE, FALSE)
)

}
