% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_keys}
\alias{aggregate_keys}
\title{Expand a dataset to include all levels of aggregation}
\usage{
aggregate_keys(data, structure, ...)
}
\arguments{
\item{data}{A dataset to aggregate.}

\item{structure}{The specification of aggregation structure.}

\item{...}{Further arguments passed to other methods.}
}
\description{
All keys must be of the same length over the same time period, and a grouping
aggregation structure will be used.
}
\examples{
library(tsibble)
aggregate_keys(tourism)

}
