% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregation-vctrs}
\alias{aggregation-vctrs}
\alias{vec_ptype2.agg_vec}
\alias{vec_ptype2.agg_vec.agg_vec}
\alias{vec_ptype2.agg_vec.default}
\alias{vec_ptype2.agg_vec.character}
\alias{vec_ptype2.character.agg_vec}
\alias{vec_ptype_abbr.agg_vec}
\alias{vec_cast.agg_vec}
\alias{vec_cast.agg_vec.agg_vec}
\alias{vec_cast.agg_vec.default}
\alias{vec_cast.character.agg_vec}
\alias{vec_proxy_compare.agg_vec}
\title{Internal vctrs methods}
\usage{
vec_ptype2.agg_vec(x, y, ...)

\method{vec_ptype2}{agg_vec.agg_vec}(x, y, ...)

\method{vec_ptype2}{agg_vec.default}(x, y, ...)

\method{vec_ptype2}{agg_vec.character}(x, y, ...)

\method{vec_ptype2}{character.agg_vec}(x, y, ...)

\method{vec_ptype_abbr}{agg_vec}(x, ...)

vec_cast.agg_vec(x, to, ...)

\method{vec_cast}{agg_vec.agg_vec}(x, to, ...)

\method{vec_cast}{agg_vec.default}(x, to, ...)

\method{vec_cast}{character.agg_vec}(x, to, ...)

\method{vec_proxy_compare}{agg_vec}(x, ...)
}
\description{
These methods are the extensions that allow aggregation vectors to work with
vctrs.
}
\keyword{internal}
