% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherent.R
\name{as_coherent_tsibble}
\alias{as_coherent_tsibble}
\alias{as_coherent_tsibble.tbl_ts}
\alias{as_coherent_tsibble.grouped_ts}
\alias{as_coherent_tsibble.tbl_df}
\alias{as_coherent_tsibble.str_tbl_ts}
\alias{as_coherent_tsibble.grouped_df}
\title{Coerce to a coherent_tsibble object}
\usage{
as_coherent_tsibble(x, ...)

\method{as_coherent_tsibble}{tbl_ts}(x, structure, ...)

\method{as_coherent_tsibble}{grouped_ts}(x, structure, ...)

\method{as_coherent_tsibble}{tbl_df}(x, ...)

\method{as_coherent_tsibble}{str_tbl_ts}(x, structure, ...)

\method{as_coherent_tsibble}{grouped_df}(x, ...)
}
\arguments{
\item{x}{Object to be coerced to a coherent_tsibble (\code{str_tbl_ts})}

\item{...}{Additional arguments passed to methods}
}
\description{
Coerce to a coherent_tsibble object
}
