% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable.R
\name{as_fable}
\alias{as_fable}
\alias{as_fable.tbl_ts}
\alias{as_fable.grouped_ts}
\alias{as_fable.tbl_df}
\alias{as_fable.fbl_ts}
\alias{as_fable.grouped_df}
\title{Coerce to a fable object}
\usage{
as_fable(x, ...)

\method{as_fable}{tbl_ts}(x, response, distribution, ...)

\method{as_fable}{grouped_ts}(x, response, distribution, ...)

\method{as_fable}{tbl_df}(x, response, distribution, ...)

\method{as_fable}{fbl_ts}(x, response, distribution, ...)

\method{as_fable}{grouped_df}(x, response, distribution, ...)
}
\arguments{
\item{x}{Object to be coerced to a fable (\code{fbl_ts})}

\item{...}{Additional arguments passed to methods}

\item{response}{The character vector of response variable(s).}

\item{distribution}{The name of the distribution column (can be provided
using a bare expression).}
}
\description{
Coerce to a fable object
}
