% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_combination.R
\name{cmbn_ensemble}
\alias{cmbn_ensemble}
\title{Ensemble combination}
\usage{
cmbn_ensemble(...)
}
\arguments{
\item{...}{Estimated models used in the ensemble.}
}
\description{
Ensemble combination
}
