% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combination.R
\name{cmbn_model}
\alias{cmbn_model}
\title{Combination modelling}
\usage{
cmbn_model(..., cmbn_fn = cmbn_ensemble)
}
\arguments{
\item{...}{Model definitions used in the combination (such as \code{\link[fable:ETS]{fable::ETS()}})}

\item{cmbn_fn}{A function used to produce the combination}
}
\description{
Combination modelling
}
