% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherent.R
\name{coherent_tsibble-vctrs}
\alias{coherent_tsibble-vctrs}
\alias{vec_ptype2.str_tbl_ts}
\alias{vec_cast.str_tbl_ts}
\title{Internal vctrs methods}
\usage{
\method{vec_ptype2}{str_tbl_ts}(x, y, ...)

\method{vec_cast}{str_tbl_ts}(x, to, ...)
}
\description{
These methods are the extensions that allow coherent_tsibble objects to
work with vctrs.
}
\keyword{internal}
