% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherent.R
\name{coherent_tsibble}
\alias{coherent_tsibble}
\title{Create a coherent_tsibble object}
\usage{
coherent_tsibble(..., structure)
}
\arguments{
\item{...}{Arguments passed to \code{\link[tsibble:tsibble]{tsibble::tsibble()}}.}

\item{response}{The character vector of response variable(s).}

\item{distribution}{The name of the distribution column (can be provided
using a bare expression).}
}
\description{
A coherent_tsibble (forecast table) data class (\code{str_tbl_ts}) which is a tsibble-like data
structure for representing forecasts. In extension to the key and index from
the tsibble (\code{tbl_ts}) class, a coherent_tsibble (\code{str_tbl_ts}) must also contain a single
distribution column that uses values from the distributional package.
}
