% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_combination.R
\name{combination_ensemble}
\alias{combination_ensemble}
\title{Ensemble combination}
\usage{
combination_ensemble(...)
}
\arguments{
\item{...}{Estimated models used in the ensemble.}
}
\description{
Ensemble combination
}
