% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_combination.R
\name{combination_model}
\alias{combination_model}
\title{Combination modelling}
\usage{
combination_model(..., cmbn_fn = combination_ensemble)
}
\arguments{
\item{...}{Model definitions used in the combination (such as \code{\link[fable:ETS]{fable::ETS()}})}

\item{cmbn_fn}{A function used to produce the combination}
}
\description{
Combination modelling
}
