% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_combination.R
\name{combination_weighted}
\alias{combination_weighted}
\title{Weighted combination}
\usage{
combination_weighted(..., weights = NULL)
}
\arguments{
\item{...}{Estimated models used in the ensemble.}

\item{weights}{The numeric weights applied to each model in \code{...}}
}
\description{
Weighted combination
}
\seealso{
\code{\link[=combination_ensemble]{combination_ensemble()}}
}
