% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{construct_fc}
\alias{construct_fc}
\title{Construct a new set of forecasts}
\usage{
construct_fc(newdata, point, se, dist)
}
\arguments{
\item{newdata}{The newdata provided to the forecast function}

\item{point}{The point transformed forecasts}

\item{se}{The standard error of the transformed forecasts}

\item{dist}{The forecast distribution (typically produced using \code{new_fcdist})}
}
\description{
Allows extension packages to create a tsibble forecast object.
This structure is a suitable output for model extension package's forecast generic.
}
\details{
Backtransformations are automatically handled, and so no transformations should be specified here.
}
