% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dable.R
\name{dable}
\alias{dable}
\title{Create a new dable}
\usage{
dable(data, decomposition, parsed_model)
}
\arguments{
\item{data}{A tsibble used for decomposition}

\item{decomposition}{The decomposition table}

\item{parsed_model}{The parsed decomposition from \code{parse_model}}
}
\description{
Create a new dable
}
