% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R
\name{dcmp_model}
\alias{dcmp_model}
\title{Decomposition modelling}
\usage{
dcmp_model(dcmp, formula, ..., dcmp_args = list())
}
\arguments{
\item{dcmp}{The decomposition function (such as \code{STL})}

\item{formula}{The formula used to describe the decomposition}

\item{...}{Model definitions used to model the components (such as \code{ETS})}

\item{dcmp_args}{Arguments to be passed to the decomposition function (\code{.f})}
}
\description{
Decomposition modelling
}
