% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R
\name{decompose}
\alias{decompose}
\alias{decompose.tbl_ts}
\title{Model based decomposition}
\usage{
decompose(.data, ...)

\method{decompose}{tbl_ts}(.data, .model, ...)
}
\arguments{
\item{.data}{A data structure suitable for the models (such as a \code{tsibble}).}

\item{...}{Further arguments passed to methods.}

\item{.model}{Definition for the model to be used.}
}
\description{
Model based decomposition
}
