% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{define_model}
\alias{define_model}
\title{Define a model}
\usage{
define_model(train, specials)
}
\arguments{
\item{train}{A function used to train the model to data}

\item{specials}{A list of functions used to be evaluated from the formula.
If specials is NULL, no specials are computed}
}
\description{
Define a model
}
