% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{model_definition}
\alias{model_definition}
\alias{decomposition_definition}
\title{R6 Objects}
\description{
Suitable for inheritance when creating new models or definitions.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-check}{\code{model_definition$check()}}
\item \href{#method-prepare}{\code{model_definition$prepare()}}
\item \href{#method-new}{\code{model_definition$new()}}
\item \href{#method-recall_lag}{\code{model_definition$recall_lag()}}
\item \href{#method-add_data}{\code{model_definition$add_data()}}
\item \href{#method-remove_data}{\code{model_definition$remove_data()}}
\item \href{#method-train}{\code{model_definition$train()}}
\item \href{#method-print}{\code{model_definition$print()}}
\item \href{#method-clone}{\code{model_definition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check"></a>}}
\subsection{Method \code{check()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model_definition$check(.data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prepare"></a>}}
\subsection{Method \code{prepare()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model_definition$prepare(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model_definition$new(formula, ..., .env)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recall_lag"></a>}}
\subsection{Method \code{recall_lag()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model_definition$recall_lag(x, n = 1L, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_data"></a>}}
\subsection{Method \code{add_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model_definition$add_data(.data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_data"></a>}}
\subsection{Method \code{remove_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model_definition$remove_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\subsection{Method \code{train()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model_definition$train(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model_definition$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{model_definition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[fabletools:NA]{fabletools::NA}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-train}{\code{decomposition_definition$train()}}
\item \href{#method-print}{\code{decomposition_definition$print()}}
\item \href{#method-clone}{\code{decomposition_definition$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="fabletools" data-topic="NA" data-id="add_data">}\href{../../fabletools/html/NA.html#method-add_data}{\code{fabletools::NA$add_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fabletools" data-topic="NA" data-id="check">}\href{../../fabletools/html/NA.html#method-check}{\code{fabletools::NA$check()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fabletools" data-topic="NA" data-id="initialize">}\href{../../fabletools/html/NA.html#method-initialize}{\code{fabletools::NA$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fabletools" data-topic="NA" data-id="prepare">}\href{../../fabletools/html/NA.html#method-prepare}{\code{fabletools::NA$prepare()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fabletools" data-topic="NA" data-id="recall_lag">}\href{../../fabletools/html/NA.html#method-recall_lag}{\code{fabletools::NA$recall_lag()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="fabletools" data-topic="NA" data-id="remove_data">}\href{../../fabletools/html/NA.html#method-remove_data}{\code{fabletools::NA$remove_data()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\subsection{Method \code{train()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{decomposition_definition$train(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{decomposition_definition$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{decomposition_definition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
