% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\docType{data}
\name{directional_accuracy_measures}
\alias{directional_accuracy_measures}
\title{Directional accuracy measures}
\format{
An object of class \code{list} of length 3.
}
\usage{
directional_accuracy_measures
}
\arguments{
\item{.resid}{A vector of residuals from either the training (model accuracy)
or test (forecast accuracy) data.}

\item{.actual}{A vector of responses matching the fitted values
(for forecast accuracy, \code{new_data} must be provided).}

\item{na.rm}{Remove the missing values before calculating the accuracy measure}

\item{...}{Additional arguments for each measure.}
}
\description{
Directional accuracy measures
}
\keyword{datasets}
