% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{dist-accuracy-measures}
\alias{dist-accuracy-measures}
\alias{percentile_score}
\title{Distribution accuracy measures}
\usage{
percentile_score(.dist, .actual, na.rm = TRUE, ...)
}
\arguments{
\item{.dist}{The distribution of fitted values from the model, or forecasted values from the forecast.}

\item{.actual}{A vector of responses matching the fitted values
(for forecast accuracy, \code{new_data} must be provided).}

\item{na.rm}{Remove the missing values before calculating the accuracy measure}

\item{...}{Additional arguments for each measure.}
}
\description{
Distribution accuracy measures
}
