% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{distribution_var}
\alias{distribution_var}
\title{Return distribution variable}
\usage{
distribution_var(x)
}
\arguments{
\item{x}{A dataset containing a distribution variable (such as a fable).}
}
\description{
\code{distribution_var()} returns a character vector of the distribution variable
in the data.
}
