% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{dist_normal}
\alias{dist_normal}
\alias{dist_sim}
\title{Distributions for intervals}
\usage{
dist_normal(mean, sd, ...)

dist_sim(sample, ...)
}
\arguments{
\item{mean}{vector of distributional means.}

\item{sd}{vector of distributional standard deviations.}

\item{...}{Additional arguments passed on to quantile methods.}

\item{sample}{a list of simulated values}
}
\description{
Distributions for intervals
}
