% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Estimate a model}
\usage{
estimate(.data, .model)
}
\arguments{
\item{.data}{A data structure suitable for the models (such as a \code{tsibble})}

\item{.model}{Definition for the model to be used}
}
\description{
Estimate a model
}
