% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconciliation.R
\name{expand_keys}
\alias{expand_keys}
\title{Expand a dataset to include all levels of aggregation}
\usage{
expand_keys(data, ...)
}
\arguments{
\item{data}{A dataset to aggregate.}

\item{...}{Further arguments passed to other methods.}
}
\description{
All keys must be of the same length over the same time period, and a grouping
aggregation structure will be used.
}
\examples{
library(tsibble)
expand_keys(tourism)

}
