% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-fable.R
\name{fable-vctrs}
\alias{fable-vctrs}
\alias{vec_ptype2.fbl_ts}
\alias{vec_cast.fbl_ts}
\title{Internal vctrs methods}
\usage{
vec_ptype2.fbl_ts(x, y, ...)

vec_cast.fbl_ts(x, to, ...)
}
\description{
These methods are the extensions that allow fable objects to
work with vctrs.
}
\keyword{internal}
