% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable.R
\name{fable}
\alias{fable}
\title{Create a fable object}
\usage{
fable(..., resp, dist)
}
\arguments{
\item{...}{Arguments passed to \code{\link[tsibble:tsibble]{tsibble::tsibble()}}.}

\item{resp}{The response variable (a list of expressions).}

\item{dist}{The distribution variable (given as a bare or unquoted variable).}
}
\description{
A fable (forecast table) data class (\code{fbl_ts}) which is a tsibble-like data
structure for representing forecasts. In extension to the key and index from
the tsibble (\code{tbl_ts}) class, a fable (\code{fbl_ts}) must contain columns of
point forecasts for the response variable(s), and a single distribution
column (\code{fcdist}).
}
