% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fablelite.R
\docType{package}
\name{fablelite-package}
\alias{fablelite}
\alias{fablelite-package}
\title{fablelite: Core tools for packages in the 'fable' framework}
\description{
Provides tools, helpers and data structures for developing models and time series functions for the 'fable' framework. These tools support a consistent and tidy interface for time series modelling and analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://fablelite.tidyverts.org/}
  \item Report bugs at \url{https://github.com/tidyverts/fablelite/issues}
}

}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com}

Authors:
\itemize{
  \item Rob Hyndman
  \item Earo Wang
}

}
\keyword{package}
