% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabletools.R
\docType{package}
\name{fabletools-package}
\alias{fabletools}
\alias{fabletools-package}
\title{fabletools: Core tools for packages in the 'fable' framework}
\description{
Provides tools, helpers and data structures for developing models and time series functions for 'fable' and extension packages. These tools support a consistent and tidy interface for time series modelling and analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://fabletools.tidyverts.org/}
  \item Report bugs at \url{https://github.com/tidyverts/fabletools/issues}
}

}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com}

Authors:
\itemize{
  \item Rob Hyndman
  \item Earo Wang
}

Other contributors:
\itemize{
  \item Di Cook [contributor]
}

}
\keyword{package}
