% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{new_fcdist}
\alias{new_fcdist}
\alias{new_fcdist_env}
\title{Create a forecast distribution object}
\usage{
new_fcdist(..., .env)

new_fcdist_env(quantile, transformation = identity, display = NULL)
}
\arguments{
\item{...}{Arguments for \code{f} function}

\item{.env}{An environment produced using \code{new_fcdist_env}}

\item{quantile}{A distribution function producing quantiles (such as \code{qnorm})}

\item{transformation}{Transformation to be applied to resulting quantiles
from \code{quantile}}

\item{display}{Function that is used to format the distribution display}
}
\description{
Create a forecast distribution object
}
\examples{
tnorm <- new_fcdist_env(qnorm, exp)
mydist <- new_fcdist(qnorm, transformation = exp,
 mean = rep(3, 10), sd = seq(0, 1, length.out=10))
mydist
hilo(mydist, 95)

}
