% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{feature_set}
\alias{feature_set}
\title{Create a feature set from tags}
\usage{
feature_set(package, tags)
}
\arguments{
\item{package}{The package(s) from which to search for features}

\item{tags}{Identifying tags}
}
\description{
Create a feature set from tags
}
