% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{features_by_pkg}
\alias{features_by_pkg}
\title{Features by package}
\description{
This documentation lists all available in currently loaded packages. This is
a useful reference for making a \code{\link[=feature_set]{feature_set()}} from particular package(s).
}
\details{
\Sexpr[stage=render,results=rd]{fabletools:::rd_features_pkg()}
}
\seealso{
\link{features_by_tag}
}
\keyword{internal}
