% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.mdl_df}
\alias{fitted.mdl_df}
\alias{fitted.mdl_ts}
\title{Extract fitted values from models}
\usage{
\method{fitted}{mdl_df}(object, ...)

\method{fitted}{mdl_ts}(object, ...)
}
\arguments{
\item{object}{A mable or time series model.}

\item{...}{Other arguments passed to the model method for \code{fitted()}}
}
\description{
Extracts the fitted values from each of the models in a mable. A tsibble will
be returned containing these fitted values. Fitted values will be
automatically back-transformed if a transformation was specified.
}
