% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast.mdl_df}
\alias{forecast.mdl_df}
\title{Forecast a mable}
\usage{
\method{forecast}{mdl_df}(object, new_data = NULL, h = NULL,
  bias_adjust = TRUE, ...)
}
\arguments{
\item{bias_adjust}{Use adjusted back-transformed mean for transformations.
Refer to \code{vignette("transformations")} for more details.}
}
\description{
Forecast a mable
}
