% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate.mdl_df}
\alias{generate.mdl_df}
\title{Generate responses from a mable}
\usage{
\method{generate}{mdl_df}(x, new_data = NULL, h = NULL, times = 1,
  seed = NULL, ...)
}
\arguments{
\item{new_data}{The data to be generated (time index and exogenous regressors)}

\item{h}{The simulation horizon (can be used instead of \code{new_data} for regular
time series with no exogenous regressors).}

\item{times}{The number of replications}

\item{seed}{The seed for the random generation from distributions}
}
\description{
Use a model's fitted distribution to simulate additional data with similar
behaviour to the response. This is a tidy implementation of
\code{\link[stats]{simulate}}.
}
\examples{
library(dplyr)
library(fable)
UKLungDeaths <- as_tsibble(cbind(mdeaths, fdeaths), pivot_longer = FALSE)
UKLungDeaths \%>\% 
  model(lm = TSLM(mdeaths ~ fourier("year", K = 4) + fdeaths)) \%>\% 
  generate(UKLungDeaths, times = 5)

}
