% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\docType{data}
\name{interval-accuracy-measures}
\alias{interval-accuracy-measures}
\alias{winkler_score}
\alias{interval_measures}
\title{Interval estimate accuracy measures}
\format{An object of class \code{list} of length 1.}
\usage{
winkler_score(.dist, .actual, level, na.rm = TRUE)

interval_measures
}
\arguments{
\item{.dist}{The distribution of fitted values from the model, or forecasted values from the forecast.}

\item{.actual}{A vector of responses matching the fitted values
(for forecast accuracy, \code{new_data} must be provided).}

\item{na.rm}{Remove the missing values before calculating the accuracy measure}
}
\description{
Interval estimate accuracy measures
}
\keyword{datasets}
