% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{is_aggregated}
\alias{is_aggregated}
\title{Is the element an aggregation of smaller data}
\usage{
is_aggregated(x)
}
\arguments{
\item{x}{An object.}
}
\description{
Is the element an aggregation of smaller data
}
\seealso{
\code{\link{aggregate_key}}
}
