% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherent.R
\name{is_coherent_tsibble}
\alias{is_coherent_tsibble}
\title{Is the object a coherent_tsibble}
\usage{
is_coherent_tsibble(x)
}
\arguments{
\item{x}{An object.}
}
\description{
Is the object a coherent_tsibble
}
