% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{MAAPE}
\alias{MAAPE}
\title{Mean Arctangent Absolute Percentage Error}
\usage{
MAAPE(.resid, .actual, na.rm = TRUE, ...)
}
\arguments{
\item{.resid}{A vector of residuals from either the training (model accuracy)
or test (forecast accuracy) data.}

\item{.actual}{A vector of responses matching the fitted values
(for forecast accuracy, \code{new_data} must be provided).}

\item{na.rm}{Remove the missing values before calculating the accuracy measure}

\item{...}{Additional arguments for each measure.}
}
\description{
Mean Arctangent Absolute Percentage Error
}
\references{
Kim, Sungil and Heeyoung Kim (2016) "A new metric of absolute percentage error
for intermittent demand forecasts". \emph{International Journal of Forecasting},
\bold{32}(3), 669-679.
}
