% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_vars.R
\name{mable_models}
\alias{mable_models}
\title{Extracts the names of the models in a model table}
\usage{
mable_models(x)

mable_models(x)
}
\arguments{
\item{x}{An object of class \code{mdl_df}.}
}
\value{
A character string.
}
\description{
Extracts the names of the models in a model table
}
