% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{mable_vars}
\alias{mable_vars}
\title{Return model column variables}
\usage{
mable_vars(x)
}
\arguments{
\item{x}{A dataset containing models (such as a mable).}
}
\description{
\code{mable_vars()} returns a character vector of the model variables in the
object.
}
