% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{MDPV}
\alias{MDPV}
\title{Mean Directional Percentage Value}
\usage{
MDPV(.resid, .actual, na.rm = TRUE, ...)
}
\arguments{
\item{.resid}{A vector of residuals from either the training (model accuracy)
or test (forecast accuracy) data.}

\item{.actual}{A vector of responses matching the fitted values
(for forecast accuracy, \code{new_data} must be provided).}

\item{na.rm}{Remove the missing values before calculating the accuracy measure}

\item{...}{Additional arguments for each measure.}
}
\description{
Mean Directional Percentage Value
}
\references{
Blaskowitz and H. Herwartz (2011) "On economic evaluation of directional forecasts". \emph{International Journal of Forecasting},
\bold{27}(4), 1058-1065.
}
