% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\docType{data}
\name{model_definition}
\alias{model_definition}
\title{Model definition}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
model_definition
}
\description{
Model definition
}
\keyword{datasets}
