% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{model_vars}
\alias{model_vars}
\title{Return model column variables}
\usage{
model_vars(x)
}
\arguments{
\item{x}{A dataset containing models (such as a mable).}
}
\description{
\code{model_vars()} returns a character vector of the model variables in the
object.
}
