% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dable.R
\name{new_dable}
\alias{new_dable}
\title{Constructor}
\usage{
new_dable(x)
}
\arguments{
\item{x}{A dable-like object}
}
\description{
A constructor function for producing a dable (most useful for extension package authors)
}
