% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{new_decomposition_definition}
\alias{new_decomposition_definition}
\title{Helper to create a new decomposition function}
\usage{
new_decomposition_definition(.class, .data, ..., .env = caller_env(n =
  2))
}
\arguments{
\item{.class}{A decomposition class (typically created with \code{\link[=new_decomposition_class]{new_decomposition_class()}}).}

\item{.data}{A tsibble.}

\item{...}{The user inputs, such as the formula and any control parameters.}

\item{.env}{The environment from which the user's objects can be found.}
}
\description{
Helper to create a new decomposition function
}
