% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.R
\name{new_model_specification}
\alias{new_model_specification}
\title{Functions associated with specifying a model}
\usage{
new_model_specification(formula, ..., class = NULL)
}
\description{
Functions associated with specifying a model
}
