% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_model_lhs}
\alias{parse_model_lhs}
\title{Parse the RHS of the model formula for transformations}
\usage{
parse_model_lhs(model, data)
}
\arguments{
\item{data}{Data to be used to find the response}

\item{model_lhs}{The expression for the lhs of the model (from \code{model_lhs()})}
}
\description{
Parse the RHS of the model formula for transformations
}
