% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_model_rhs}
\alias{parse_model_rhs}
\title{Parse the RHS of the model formula for specials}
\usage{
parse_model_rhs(model)
}
\arguments{
\item{model}{A model definition}
}
\description{
Parse the RHS of the model formula for specials
}
\keyword{internal}
