% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{ME}
\alias{ME}
\alias{MSE}
\alias{RMSE}
\alias{MAE}
\alias{MPE}
\alias{MAPE}
\alias{MASE}
\alias{ACF1}
\title{Point estimate accuracy measures}
\usage{
ME(.resid, na.rm = TRUE, ...)

MSE(.resid, na.rm = TRUE, ...)

RMSE(.resid, na.rm = TRUE, ...)

MAE(.resid, na.rm = TRUE, ...)

MPE(.resid, .resp, na.rm = TRUE, ...)

MAPE(.resid, .resp, na.rm = TRUE, ...)

MASE(.resid, .resp, demean = FALSE, na.rm = TRUE, .period,
  d = .period == 1, D = .period > 1, ...)

ACF1(.resid, na.action = stats::na.pass, ...)
}
\arguments{
\item{.resid}{A vector of residuals from either the training (model accuracy) or test (forecast accuracy) data.}

\item{na.rm}{Remove the missing values before calculating the accuracy measure}

\item{...}{Additional arguments for each measure}

\item{.resp}{A vector of responses matching the residuals (for forecast accuracy, the original data must be provided).}

\item{demean}{Should the response be demeaned (MASE)}

\item{.period}{The seasonal period of the data (defaulting to 'smallest' seasonal period).}

\item{d}{Should the response model include a first difference?}

\item{D}{Should the response model include a seasonal difference?}
}
\description{
Point estimate accuracy measures
}
