% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconciliation.R
\name{reconcile}
\alias{reconcile}
\alias{reconcile.mbl_df}
\title{Forecast reconciliation}
\usage{
reconcile(.data, ...)

\method{reconcile}{mbl_df}(.data, ...)
}
\arguments{
\item{.data}{A mable.}

\item{...}{Reconciliation methods applied to model columns within \code{.data}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
\code{reconcile()} is deprecated in favour of \code{\link[dplyr:mutate]{dplyr::mutate()}} with the
\verb{reconcile_*()} functions (\code{\link[=reconcile_bu]{reconcile_bu()}}, \code{\link[=reconcile_td]{reconcile_td()}},
\code{\link[=reconcile_mo]{reconcile_mo()}} and \code{\link[=reconcile_mint]{reconcile_mint()}}).
}
\examples{
\dontshow{if (requireNamespace("fable", quietly = TRUE)) withAutoprint(\{ # examplesIf}
library(fable)
lung_deaths_agg <- as_tsibble(cbind(mdeaths, fdeaths)) \%>\%
  aggregate_key(key, value = sum(value))

lung_deaths_agg \%>\%
  model(lm = TSLM(value ~ trend() + season())) \%>\%
  dplyr::mutate(lm = reconcile_mint(lm)) \%>\% 
  forecast()
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=aggregate_key]{aggregate_key()}}

\href{https://otexts.com/fpp3/hierarchical.html}{\emph{Forecasting: Principles and Practice} - Forecasting hierarchical and grouped time series}
}
