% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{as_tsibble}
\alias{vars}
\alias{autoplot}
\alias{autolayer}
\alias{accuracy}
\alias{equation}
\alias{interpolate}
\alias{components}
\alias{augment}
\alias{glance}
\alias{tidy}
\alias{hypothesize}
\alias{generate}
\alias{refit}
\alias{forecast}
\alias{hilo}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{distributional}{\code{\link[distributional:hilo]{hilo()}}}

  \item{dplyr}{\code{\link[dplyr:\%>\%]{\%>\%}}, \code{\link[dplyr:vars]{vars()}}}

  \item{generics}{\code{\link[generics:accuracy]{accuracy()}}, \code{\link[generics:augment]{augment()}}, \code{\link[generics:components]{components()}}, \code{\link[generics:equation]{equation()}}, \code{\link[generics:forecast]{forecast()}}, \code{\link[generics:generate]{generate()}}, \code{\link[generics:glance]{glance()}}, \code{\link[generics:hypothesize]{hypothesize()}}, \code{\link[generics:interpolate]{interpolate()}}, \code{\link[generics:refit]{refit()}}, \code{\link[generics:tidy]{tidy()}}}

  \item{ggplot2}{\code{\link[ggplot2:autolayer]{autolayer()}}, \code{\link[ggplot2:autoplot]{autoplot()}}}

  \item{tsibble}{\code{\link[tsibble:as_tsibble]{as_tsibble()}}}
}}

