% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit.R
\name{refit}
\alias{refit}
\title{Refit a model with new data}
\usage{
refit(object, new_data, ...)
}
\arguments{
\item{object}{A model (typically a mable)}

\item{new_data}{A tsibble containing the new data}

\item{...}{Additional optional arguments for refit methods}
}
\description{
Refit a model with new data
}
