% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{register_feature}
\alias{register_feature}
\title{Register a feature function}
\usage{
register_feature(fn, tags)
}
\arguments{
\item{fn}{The feature function}

\item{tags}{Identifying tags}
}
\description{
Allows users to find and use features from your package using \code{\link[=feature_set]{feature_set()}}.
}
