% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.mdl_df}
\alias{residuals.mdl_df}
\alias{residuals.mdl_ts}
\title{Extract residuals values from models}
\usage{
\method{residuals}{mdl_df}(object, ...)

\method{residuals}{mdl_ts}(object, type = "innovation", ...)
}
\arguments{
\item{object}{A mable or time series model.}

\item{...}{Other arguments passed to the model method for \code{residuals()}}

\item{type}{The type of residuals to compute. If \code{type="response"}, residuals on the back-transformed data will be computed.}
}
\description{
Extracts the residuals from each of the models in a mable. A tsibble will
be returned containing these residuals.
}
