% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{response}
\alias{response}
\title{Extract the respone data from a model}
\usage{
response(object, ...)
}
\arguments{
\item{object}{The object containing response data}

\item{...}{Additional parameters passed on to other methods}
}
\description{
Extract the respone data from a model
}
