% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_vars.R
\name{response_var}
\alias{response_var}
\alias{response_var.fbl_ts}
\alias{response_var.mdl_df}
\alias{response_var.mdl_ts}
\title{Extracts the response variable for a model, model table or forecast table}
\usage{
response_var(x)

\method{response_var}{fbl_ts}(x)

\method{response_var}{mdl_df}(x)

\method{response_var}{mdl_ts}(x)
}
\arguments{
\item{x}{An object of class \code{mdl_ts}, \code{mdl_df} or \code{fbl_ts}.}
}
\value{
A character string.
}
\description{
This function returns the name of the response variable in a model. For a forecast table, it returns the name of the response variable in the model(s) used to obtain the forecasts.
}
