% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{response_vars}
\alias{response_vars}
\title{Return response variables}
\usage{
response_vars(x)
}
\arguments{
\item{x}{A dataset containing a response variable (such as a mable, fable, or dable).}
}
\description{
\code{response_vars()} returns a character vector of the response variables in the
object.
}
