% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{sample_quantile}
\alias{sample_quantile}
\title{Prepare a forecast distribution from simulated cases}
\usage{
sample_quantile(x, ...)
}
\arguments{
\item{x}{A list of sampled quantities}

\item{...}{Additional arguments for \code{quantile}}
}
\description{
Prepare a forecast distribution from simulated cases
}
