% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{scenarios}
\alias{scenarios}
\title{A set of future scenarios for forecasting}
\usage{
scenarios(..., names_to = ".scenario")
}
\arguments{
\item{...}{Input data for each scenario}

\item{names_to}{The column name used to identify each scenario}
}
\description{
A set of future scenarios for forecasting
}
