% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specials.R
\name{special_xreg}
\alias{special_xreg}
\title{Special for producing a model matrix of exogenous regressors}
\usage{
special_xreg(...)
}
\arguments{
\item{...}{Bare expressions to be used as regressors in the model}
}
\description{
Special for producing a model matrix of exogenous regressors
}
