% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_specials.R
\name{specials_fn_bind}
\alias{specials_fn_bind}
\title{Bind values to the functions of a specials environment}
\usage{
specials_fn_bind(.specials, ...)
}
\arguments{
\item{.specials}{The specials environment}

\item{...}{Values to be bound to the functions of the specials environment}
}
\description{
Bind values to the functions of a specials environment
}
