% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.R
\name{tidy.mbl_df}
\alias{tidy.mbl_df}
\alias{coef.mbl_df}
\alias{coef.mdl_lst}
\alias{coef.mdl_ts}
\alias{tidy.mdl_lst}
\alias{tidy.mdl_ts}
\title{Extract model coefficients from a mable}
\usage{
\method{tidy}{mbl_df}(x, ...)

\method{coef}{mbl_df}(object, ...)

\method{coef}{mdl_lst}(object, ...)

\method{coef}{mdl_ts}(object, ...)

\method{tidy}{mdl_lst}(x, ...)

\method{tidy}{mdl_ts}(x, ...)
}
\arguments{
\item{x, object}{A mable.}

\item{...}{Arguments for model methods.}
}
\description{
This function will obtain the coefficients (and associated statistics) for
each model in the mable.
}
\examples{
\dontshow{if (requireNamespace("fable", quietly = TRUE)) withAutoprint(\{ # examplesIf}
library(fable)
library(tsibbledata)

olympic_running \%>\%
  model(lm = TSLM(log(Time) ~ trend())) \%>\% 
  tidy()
\dontshow{\}) # examplesIf}
}
