% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconciliation.R
\name{reconcile_td}
\alias{reconcile_td}
\alias{top_down}
\title{Top down forecast reconciliation}
\usage{
reconcile_td(
  models,
  method = c("forecast_proportions", "average_proportions", "proportion_averages")
)

top_down(
  models,
  method = c("forecast_proportions", "average_proportions", "proportion_averages")
)
}
\arguments{
\item{models}{A column of models in a mable.}

\item{method}{The method used to compute the disaggregation proportions.}
}
\description{
\lifecycle{experimental}
}
\details{
Reconciles a hierarchy using the top down reconciliation method. The
response variable of the hierarchy must be aggregated using sums. The
forecasted time points must match for all series in the hierarchy.

\code{reconcile_td()} is a short alias for \code{top_down()}.
}
\seealso{
\code{\link[=reconcile_bu]{reconcile_bu()}}, \code{\link[=reconcile_mo]{reconcile_mo()}}, \code{\link[=reconcile_mint]{reconcile_mint()}}, \code{\link[=aggregate_key]{aggregate_key()}}

\href{https://otexts.com/fpp3/single-level.html#top-down-approaches}{\emph{Forecasting: Principles and Practice} - Top-down approaches}
}
