% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transformation}
\alias{transformation}
\title{Extract a transformation from an object}
\usage{
transformation(x, ...)
}
\arguments{
\item{x}{An object containing a transformation}

\item{...}{Additional arguments passed on to other methods}
}
\description{
Extract a transformation from an object
}
