% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{validate_formula}
\alias{validate_formula}
\title{Validate the user provided model}
\usage{
validate_formula(model, data = NULL)
}
\arguments{
\item{model}{A quosure for the user's model specification}

\item{data}{A dataset used for automatic response selection}
}
\description{
Appropriately format the user's model for evaluation. Typically ran as one of the first steps
in a model function.
}
\keyword{internal}
