% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\title{Evaluate model/forecast accuracy}
\usage{
accuracy(x, ...)
}
\description{
The measures calculated are:
\itemize{
\item ME: Mean Error
\item RMSE: Root Mean Squared Error
\item MAE: Mean Absolute Error
\item MPE: Mean Percentage Error
\item MAPE: Mean Absolute Percentage Error
\item MASE: Mean Absolute Scaled Error
\item ACF1: Autocorrelation of errors at lag 1.
}
}
