% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\title{Evaluate model/forecast accuracy}
\usage{
accuracy(object, ...)
}
\arguments{
\item{object}{A model or forecast object}

\item{...}{Additional arguments to be passed to measures that use it.

The measures calculated are:
\itemize{
\item ME: Mean Error
\item RMSE: Root Mean Squared Error
\item MAE: Mean Absolute Error
\item MPE: Mean Percentage Error
\item MAPE: Mean Absolute Percentage Error
\item MASE: Mean Absolute Scaled Error
\item ACF1: Autocorrelation of errors at lag 1.
}}
}
\description{
Evaluate model/forecast accuracy
}
