% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_keys}
\alias{aggregate_keys}
\alias{aggregate_index}
\title{Expand a dataset to include other levels of aggregation}
\usage{
aggregate_keys(.data, .spec, ...)

aggregate_index(.data, .times, ...)
}
\arguments{
\item{.data}{A tsibble.}

\item{.spec}{The specification of aggregation structure.}

\item{...}{Name-value pairs of summary functions. The name will be the
name of the variable in the result. The value should be an expression
that returns a single value like \code{min(x)}, \code{n()}, or \code{sum(is.na(y))}.

The arguments in \code{...} are automatically \link[rlang:quo]{quoted} and
\link[rlang:eval_tidy]{evaluated} in the context of the data
frame. They support \link[rlang:quasiquotation]{unquoting} and
splicing. See \code{vignette("programming")} for an introduction to
these concepts.}

\item{.times}{Temporal aggregations to include. The default (NULL) will
automatically identify appropriate temporal aggregations. This can be specified
as a list of \code{\link[lubridate:period]{lubridate::period()}} elements, or a character vector describing the
temporal aggregations.}
}
\description{
Uses the structural specification given in \code{.spec} to aggregate a time
series. Commonly used in combination with forecast reconciliation.
}
\details{
A grouped structure is specified using \code{grp1 * grp2}, and a nested structure
is specified via \code{parent / child}.
}
\examples{
library(tsibble)
tourism_grp <- tourism \%>\% 
  aggregate_keys(Purpose * (State / Region), Trips = sum(Trips))

library(tsibble)
pedestrian \%>\% 
  aggregate_index()

}
