% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dable.R
\name{as_dable}
\alias{as_dable}
\alias{as_dable.tbl_ts}
\title{Coerce to a dable object}
\usage{
as_dable(x, ...)

\method{as_dable}{tbl_ts}(x, resp, dcmp, structure = list(), ...)
}
\arguments{
\item{x}{Object to be coerced to a dable (\code{dcmp_ts})}

\item{...}{Additional arguments passed to methods}

\item{resp}{The response variable (given as a bare or unquoted variable).}

\item{dcmp}{An expression of data columns defining how the variable was decomposed.}

\item{structure}{A named list describing the structure of the components
(such as \code{period}, and \code{base}).}
}
\description{
Coerce to a dable object
}
