% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dable.R
\name{as_dable}
\alias{as_dable}
\alias{as_dable.tbl_ts}
\title{Coerce to a dable object}
\usage{
as_dable(x, ...)

\method{as_dable}{tbl_ts}(x, resp, seasons = list(), aliases = list(),
  ...)
}
\arguments{
\item{x}{Object to be coerced to a dable (\code{dcmp_ts})}

\item{...}{Additional arguments passed to methods}

\item{resp}{The response variable (given as a bare or unquoted variable).}

\item{seasons}{A named list describing the structure of seasonal components
(such as \code{period}, and \code{base}).}

\item{aliases}{A named list of calls describing common aliases computed from
components.}
}
\description{
Coerce to a dable object
}
