% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dable.R
\name{as_dable}
\alias{as_dable}
\alias{as_dable.tbl_df}
\alias{as_dable.tbl_ts}
\title{Coerce to a dable object}
\usage{
as_dable(x, ...)

\method{as_dable}{tbl_df}(x, response, method = NULL, seasons = list(), aliases = list(), ...)

\method{as_dable}{tbl_ts}(x, response, method = NULL, seasons = list(), aliases = list(), ...)
}
\arguments{
\item{x}{Object to be coerced to a dable (\code{dcmp_ts})}

\item{...}{Additional arguments passed to methods}

\item{response}{The character vector of response variable(s).}

\item{method}{The name of the decomposition method.}

\item{seasons}{A named list describing the structure of seasonal components
(such as \code{period}, and \code{base}).}

\item{aliases}{A named list of calls describing common aliases computed from
components.}
}
\description{
Coerce to a dable object
}
