% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable.R
\name{as_fable}
\alias{as_fable}
\alias{as_fable.tbl_ts}
\alias{as_fable.grouped_ts}
\alias{as_fable.tbl_df}
\alias{as_fable.fbl_ts}
\alias{as_fable.grouped_df}
\title{Coerce to a fable object}
\usage{
as_fable(x, ...)

\method{as_fable}{tbl_ts}(x, resp, dist, ...)

\method{as_fable}{grouped_ts}(x, resp, dist, ...)

\method{as_fable}{tbl_df}(x, resp, dist, ...)

\method{as_fable}{fbl_ts}(x, resp, dist, ...)

\method{as_fable}{grouped_df}(x, resp, dist, ...)
}
\arguments{
\item{x}{Object to be coerced to a fable (\code{fbl_ts})}

\item{...}{Additional arguments passed to methods}

\item{resp}{The response variable (a list of expressions).}

\item{dist}{The distribution variable (given as a bare or unquoted variable).}
}
\description{
Coerce to a fable object
}
