% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable.R
\name{as_fable}
\alias{as_fable}
\alias{as_fable.tbl_ts}
\title{Coerce to a fable object}
\usage{
as_fable(x, ...)

\method{as_fable}{tbl_ts}(x, resp, dist, ...)
}
\arguments{
\item{x}{Object to be coerced to a fable (\code{fbl_ts})}

\item{...}{A set of name-value pairs. The names of "key" and "index" should
be avoided as they are used as the arguments.}

\item{resp}{The response variable (given as a bare or unquoted variable).}

\item{dist}{The distribution variable (given as a bare or unquoted variable).}
}
\description{
Coerce to a fable object
}
