% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mable.R
\name{as_mable}
\alias{as_mable}
\title{Coerce a dataset to a mable}
\usage{
as_mable(x, key = structure(exprs(), class = "key"))
}
\arguments{
\item{x}{A dataset containing a list model column}

\item{key}{Structural variable(s) that identify each model}
}
\description{
Coerce a dataset to a mable
}
