% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mable.R
\name{as_mable}
\alias{as_mable}
\alias{as_mable.tbl_df}
\title{Coerce a dataset to a mable}
\usage{
as_mable(x, ...)

\method{as_mable}{tbl_df}(x, key = NULL, models = NULL, ...)
}
\arguments{
\item{x}{A dataset containing a list model column.}

\item{...}{Additional arguments passed to other methods.}

\item{key}{Structural variable(s) that identify each model.}

\item{models}{Identifiers for the columns containing model(s).}
}
\description{
Coerce a dataset to a mable
}
