% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconciliation.R
\name{reconcile_bu}
\alias{reconcile_bu}
\alias{bottom_up}
\title{Bottom up forecast reconciliation}
\usage{
reconcile_bu(models)

bottom_up(models)
}
\arguments{
\item{models}{A column of models in a mable.}
}
\description{
\lifecycle{experimental}
}
\details{
Reconciles a hierarchy using the bottom up reconciliation method. The
response variable of the hierarchy must be aggregated using sums. The
forecasted time points must match for all series in the hierarchy.

\code{reconcile_bu()} is a short alias for \code{bottom_up()}.
}
\seealso{
\code{\link[=reconcile_td]{reconcile_td()}}, \code{\link[=reconcile_mo]{reconcile_mo()}}, \code{\link[=reconcile_mint]{reconcile_mint()}}, \code{\link[=aggregate_key]{aggregate_key()}}

\href{https://otexts.com/fpp3/single-level.html#bottom-up-approach}{\emph{Forecasting: Principles and Practice} - Bottom-up approach}
}
