% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_combination.R
\name{combination_ensemble}
\alias{combination_ensemble}
\title{Ensemble combination}
\usage{
combination_ensemble(..., weights = c("equal", "inv_var"))
}
\arguments{
\item{...}{Estimated models used in the ensemble.}

\item{weights}{The method used to weight each model in the ensemble.}
}
\description{
Ensemble combination
}
