% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_combination.R
\name{combination_model}
\alias{combination_model}
\title{Combination modelling}
\usage{
combination_model(..., cmbn_fn = combination_ensemble)
}
\arguments{
\item{...}{Model definitions used in the combination (such as \code{\link[fable:ETS]{fable::ETS()}})}

\item{cmbn_fn}{A function used to produce the combination}
}
\description{
Combines multiple model definitions (passed via \code{...}) to produce a model
combination definition using some combination function (\code{cmbn_fn}).
}
\details{
A combination model can also be produced using mathematical operations.
}
\examples{
library(fable)
library(tsibble)
library(tsibbledata)

aus_production \%>\%
  model(
    cmbn1 = combination_model(SNAIVE(Beer), TSLM(Beer ~ trend() + season())),
    cmbn2 = (SNAIVE(Beer) + TSLM(Beer ~ trend() + season()))/2
  )
  
}
