% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{construct_fc}
\alias{construct_fc}
\title{Construct a new set of forecasts}
\usage{
construct_fc(point, sd, dist)
}
\arguments{
\item{point}{The transformed point forecasts}

\item{sd}{The standard deviation of the transformed forecasts}

\item{dist}{The forecast distribution (typically produced using \code{new_fcdist})}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. \code{forecast()} methods for a model should return
a vector of distributions using the distributional package.

Backtransformations are automatically handled, and so no transformations should be specified here.
}
\keyword{internal}
