% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-dable.R
\name{dable-vctrs}
\alias{dable-vctrs}
\alias{vec_ptype2.dcmp_ts}
\alias{vec_cast.dcmp_ts}
\title{Internal vctrs methods}
\usage{
vec_ptype2.dcmp_ts(x, y, ...)

vec_cast.dcmp_ts(x, to, ...)
}
\description{
These methods are the extensions that allow dable objects to
work with vctrs.
}
\keyword{internal}
