% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dable.R
\name{dable}
\alias{dable}
\title{Create a dable object}
\usage{
dable(..., key = NULL, index, resp, method = NULL, seasons = list(),
  aliases = list(), regular = TRUE)
}
\arguments{
\item{...}{A set of name-value pairs. The names of "key" and "index" should
be avoided as they are used as the arguments.}

\item{key}{Unquoted variable(s) that uniquely determine time indices. \code{NULL} for
empty key, and works with tidy selector (e.g. \code{starts_with()}).}

\item{index}{A bare (or unquoted) variable to specify the time index variable.}

\item{resp}{The response variable (a list of expressions).}

\item{method}{The name of the decomposition method.}

\item{seasons}{A named list describing the structure of seasonal components
(such as \code{period}, and \code{base}).}

\item{aliases}{A named list of calls describing common aliases computed from
components.}

\item{regular}{Regular time interval (\code{TRUE}) or irregular (\code{FALSE}). The
interval is determined by the greatest common divisor of index column, if \code{TRUE}.}
}
\description{
Create a dable object
}
