% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R
\name{decomposition}
\alias{decomposition}
\title{Decomposition modelling}
\usage{
decomposition(.f, formula, ..., .f_args)
}
\arguments{
\item{.f}{The decomposition function (such as \code{STL})}

\item{formula}{The formula used to describe the decomposition}

\item{...}{Model definitions used to model the components (such as \code{ETS})}

\item{.f_args}{Arguments to be passed to the decomposition function (\code{.f})}
}
\description{
Decomposition modelling
}
