% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{data}
\name{model_definition}
\alias{model_definition}
\alias{decomposition_definition}
\title{R6 Objects}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
model_definition

decomposition_definition
}
\description{
Suitable for inheritance when creating new models or definitions.
}
\keyword{datasets}
