% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\docType{data}
\name{percentile_score}
\alias{percentile_score}
\alias{CRPS}
\alias{distribution_accuracy_measures}
\title{Distribution accuracy measures}
\format{
An object of class \code{list} of length 2.
}
\usage{
percentile_score(.dist, .actual, na.rm = TRUE, ...)

CRPS(.dist, .actual, n_quantiles = 1000, na.rm = TRUE, ...)

distribution_accuracy_measures
}
\arguments{
\item{.dist}{The distribution of fitted values from the model, or forecasted values from the forecast.}

\item{.actual}{A vector of responses matching the fitted values
(for forecast accuracy, \code{new_data} must be provided).}

\item{na.rm}{Remove the missing values before calculating the accuracy measure}

\item{...}{Additional arguments for each measure.}

\item{n_quantiles}{The number of quantiles to use in approximating CRPS when an exact solution is not available.}
}
\description{
Distribution accuracy measures
}
\keyword{datasets}
