% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable.R
\name{fable}
\alias{fable}
\title{Create a new fable}
\usage{
fable(mable, forecast)
}
\arguments{
\item{mable}{The mable containing models used to produce the forecasts}

\item{forecast}{A list of tsibble forecasts (from \code{construct_fc})}
}
\description{
Create a new fable
}
