% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable.R
\name{fable}
\alias{fable}
\title{Create a fable object}
\usage{
fable(..., key = id(), index, resp, dist, regular = TRUE)
}
\arguments{
\item{...}{A set of name-value pairs. The names of "key" and "index" should
be avoided as they are used as the arguments.}

\item{key}{Variable(s) that define unique time indices, used in conjunction
with the helper \code{\link[=id]{id()}}. If a univariate time series (without an explicit key),
simply call \code{id()}.}

\item{index}{A bare (or unquoted) variable to specify the time index variable.}

\item{resp}{The response variable (given as a bare or unquoted variable).}

\item{dist}{The distribution variable (given as a bare or unquoted variable).}

\item{regular}{Regular time interval (\code{TRUE}) or irregular (\code{FALSE}). The
interval is determined by the greatest common divisor of index column, if \code{TRUE}.}
}
\description{
Create a fable object
}
