% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable.R
\name{fable}
\alias{fable}
\title{Create a fable object}
\usage{
fable(..., resp, dist)
}
\arguments{
\item{...}{Arguments passed to \code{\link[tsibble:tsibble]{tsibble::tsibble()}}.}

\item{resp}{The response variable (a list of expressions).}

\item{dist}{The distribution variable (given as a bare or unquoted variable).}
}
\description{
Create a fable object
}
