% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable.R
\name{fable}
\alias{fable}
\title{Create a fable object}
\usage{
fable(..., response, distribution)
}
\arguments{
\item{...}{Arguments passed to \code{\link[tsibble:tsibble]{tsibble::tsibble()}}.}

\item{response}{The character vector of response variable(s).}

\item{distribution}{The name of the distribution column (can be provided
using a bare expression).}
}
\description{
A fable (forecast table) data class (\code{fbl_ts}) which is a tsibble-like data
structure for representing forecasts. In extension to the key and index from
the tsibble (\code{tbl_ts}) class, a fable (\code{fbl_ts}) must also contain a single
distribution column that uses values from the distributional package.
}
