% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fablelite.R
\docType{package}
\name{fablelite-package}
\alias{fablelite}
\alias{fablelite-package}
\title{fablelite: Core tools for fable extensions}
\description{
Provides tools, helpers and data structures which support models in the fable framework. Includes methods for model combination, accuracy evaluation, forecast reconciliation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidyverts/fablelite}
  \item Report bugs at \url{https://github.com/tidyverts/fablelite/issues}
}

}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com}

Authors:
\itemize{
  \item Rob Hyndman
  \item Earo Wang
}

}
\keyword{package}
