% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabletools-package.R
\docType{package}
\name{fabletools-package}
\alias{fabletools}
\alias{fabletools-package}
\title{fabletools: Core Tools for Packages in the 'fable' Framework}
\description{
Provides tools, helpers and data structures for
    developing models and time series functions for 'fable' and extension
    packages. These tools support a consistent and tidy interface for time
    series modelling and analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fabletools.tidyverts.org/}
  \item \url{https://github.com/tidyverts/fabletools}
  \item Report bugs at \url{https://github.com/tidyverts/fabletools/issues}
}

}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com}

Authors:
\itemize{
  \item Rob Hyndman
  \item Earo Wang
}

Other contributors:
\itemize{
  \item Di Cook [contributor]
  \item George Athanasopoulos [contributor]
  \item David Holt [contributor]
}

}
\keyword{package}
