% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{feature_set}
\alias{feature_set}
\title{Create a feature set from tags}
\usage{
feature_set(pkgs = NULL, tags = NULL)
}
\arguments{
\item{pkgs}{The package(s) from which to search for features. If \code{NULL},
all registered features from currently loaded packages will be searched.}

\item{tags}{Tags used to identify similar groups of features. If \code{NULL},
all tags will be included.}
}
\description{
Create a feature set from tags
}
