% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast.mdl_df}
\alias{forecast.mdl_df}
\title{Forecast a mable}
\usage{
\method{forecast}{mdl_df}(object, new_data = NULL, h = NULL,
  bias_adjust = TRUE, ...)
}
\arguments{
\item{object}{A mable containing models used for forecasting.}

\item{new_data}{A \code{tsibble} containing future information used to forecast.}

\item{h}{The forecast horison (can be used instead of \code{new_data} for regular
time series with no exogenous regressors).}

\item{bias_adjust}{Use adjusted back-transformed mean for transformations.
Refer to \code{vignette("transformations")} for more details.}

\item{...}{Further arguments to forecast model methods.}
}
\description{
Forecast a mable
}
