% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\alias{forecast.mdl_df}
\title{Produce forecasts}
\usage{
forecast(object, ...)

\method{forecast}{mdl_df}(object, new_data = NULL, h = NULL,
  bias_adjust = TRUE, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecasts}

\item{...}{Additional arguments for forecast model methods.}

\item{new_data}{A \code{tsibble} containing future information used to forecast.}

\item{h}{The forecast horison (can be used instead of \code{new_data} for regular
time series with no exogenous regressors).}

\item{bias_adjust}{Use adjusted back-transformed mean for transformations.
Refer to \code{vignette("transformations")} for more details.}
}
\description{
The forecast function allows you to produce future predictions of a time series
from fitted models.
}
