% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\title{Produce forecasts}
\usage{
forecast(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecasts}

\item{...}{Extra arguments for other methods (such as \code{new_data} and \code{h}).}
}
\description{
The forecast function allows you to produce future predictions of a time series
from fitted models.
}
