% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesise.R
\name{hypothesize.mdl_df}
\alias{hypothesize.mdl_df}
\title{Run a hypothesis test from a mable}
\usage{
\method{hypothesize}{mdl_df}(x, ...)
}
\arguments{
\item{x}{A mable.}

\item{...}{Arguments for model methods.}
}
\description{
This function will return the results of a hypothesis test for each model in
the mable.
}
\examples{
if (requireNamespace("fable", quietly = TRUE)) {
library(fable)
library(tsibbledata)

olympic_running \%>\%
  model(lm = TSLM(log(Time) ~ trend())) \%>\% 
  hypothesize()
}

}
