% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\docType{data}
\name{winkler_score}
\alias{winkler_score}
\alias{interval_accuracy_measures}
\title{Interval estimate accuracy measures}
\format{An object of class \code{list} of length 1.}
\usage{
winkler_score(.dist, .actual, level = 95, na.rm = TRUE, ...)

interval_accuracy_measures
}
\arguments{
\item{.dist}{The distribution of fitted values from the model, or forecasted values from the forecast.}

\item{.actual}{A vector of responses matching the fitted values
(for forecast accuracy, \code{new_data} must be provided).}

\item{level}{The level of the forecast interval.}

\item{na.rm}{Remove the missing values before calculating the accuracy measure}

\item{...}{Additional arguments for each measure.}
}
\description{
Interval estimate accuracy measures
}
\keyword{datasets}
