% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-mable.R
\name{mable-vctrs}
\alias{mable-vctrs}
\alias{vec_ptype2.mdl_df}
\alias{vec_cast.mdl_df}
\title{Internal vctrs methods}
\usage{
vec_ptype2.mdl_df(x, y, ...)

vec_cast.mdl_df(x, to, ...)
}
\description{
These methods are the extensions that allow mable objects to
work with vctrs.
}
\keyword{internal}
