% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconciliation.R
\name{middle_out}
\alias{middle_out}
\title{Middle out forecast reconciliation}
\usage{
middle_out(models, split = 1)
}
\arguments{
\item{models}{A column of models in a mable.}

\item{split}{The middle level of the hierarchy from which the bottom-up and
top-down approaches are used above and below respectively.}
}
\description{
\lifecycle{experimental}
}
\details{
Reconciles a hierarchy using the middle out reconciliation method. The
response variable of the hierarchy must be aggregated using sums. The
forecasted time points must match for all series in the hierarchy.
}
\seealso{
\code{\link[=reconcile]{reconcile()}}, \code{\link[=aggregate_key]{aggregate_key()}}
\href{https://otexts.com/fpp3/single-level.html#middle-out-approach}{\emph{Forecasting: Principles and Practice} - Middle-out approach}
}
