% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconciliation.R
\name{reconcile_mint}
\alias{reconcile_mint}
\alias{min_trace}
\title{Minimum trace forecast reconciliation}
\usage{
reconcile_mint(
  models,
  method = c("wls_var", "ols", "wls_struct", "mint_cov", "mint_shrink"),
  sparse = NULL
)

min_trace(
  models,
  method = c("wls_var", "ols", "wls_struct", "mint_cov", "mint_shrink"),
  sparse = NULL
)
}
\arguments{
\item{models}{A column of models in a mable.}

\item{method}{The reconciliation method to use, which determines how the
base forecast error covariance matrix is estimated.}

\item{sparse}{If TRUE, the reconciliation will be computed using sparse
matrix algebra? By default, sparse matrices will be used if the Matrix
package is installed.}
}
\description{
Reconciles a hierarchy using the minimum trace combination method. The
response variable of the hierarchy must be aggregated using sums. The
forecasted time points must match for all series in the hierarchy (caution:
this is not yet tested for beyond the series length).
}
\details{
The \code{"wls_var"}, \code{"mint_cov"} and \code{"mint_shrink"} methods require a positive
definite covariance matrix, which is not the case if two series in the
hierarchy are identical (e.g. an aggregate with only one child). Use
\code{"wls_struct"} or \code{"ols"} in these situations.

\code{reconcile_mint()} is a short alias for \code{min_trace()}.
}
\references{
Wickramasuriya, S. L., Athanasopoulos, G., & Hyndman, R. J. (2019). Optimal forecast reconciliation for hierarchical and grouped time series through trace minimization. Journal of the American Statistical Association, 1-45. https://doi.org/10.1080/01621459.2018.1448825
}
\seealso{
\code{\link[=reconcile_bu]{reconcile_bu()}}, \code{\link[=reconcile_td]{reconcile_td()}}, \code{\link[=reconcile_mo]{reconcile_mo()}}, \code{\link[=aggregate_key]{aggregate_key()}}

\href{https://otexts.com/fpp3/reconciliation.html}{\emph{Forecasting: Principles and Practice} - Forecast reconciliation}
}
