% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconciliation.R
\name{trace_min}
\alias{trace_min}
\title{Minimum trace forecast reconciliation}
\usage{
trace_min(models, method = c("shrink", "wls", "ols", "cov"),
  sparse = requireNamespace("SparseM"))
}
\arguments{
\item{models}{A column of models in a mable.}

\item{method}{The reconciliation method to use.}

\item{sparse}{Should the reconciliation be computed with sparse matrix algebra?}
}
\description{
Minimum trace forecast reconciliation
}
