% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconciliation.R
\name{MinT}
\alias{MinT}
\title{Minimum trace forecast reconciliation}
\usage{
MinT(mdls, method = "comb", weights = "ols")
}
\arguments{
\item{mdls}{A column of models in a mable}

\item{method}{The combination method}

\item{weights}{The weights used in combination}
}
\description{
Minimum trace forecast reconciliation
}
