% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconciliation.R
\name{MinT}
\alias{MinT}
\title{Minimum trace forecast reconciliation}
\usage{
MinT(mdls, method = c("WLS", "MinT_cov", "MinT_shrink"))
}
\arguments{
\item{mdls}{A column of models in a mable}

\item{method}{The reconciliation method to use.}
}
\description{
Minimum trace forecast reconciliation
}
