% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\alias{model.tbl_ts}
\title{Estimate models}
\usage{
model(.data, ...)

\method{model}{tbl_ts}(.data, ..., .safely = TRUE)
}
\arguments{
\item{.data}{A data structure suitable for the models (such as a \code{tsibble})}

\item{...}{Definitions for the models to be used}

\item{.safely}{If a model encounters an error, rather than aborting the process a \link[=null_model]{NULL model} will be returned instead. This allows for an error to occur when computing many models, without losing the results of the successful models.}
}
\description{
Trains specified model definitions to a dataset.
}
