% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\alias{model.tbl_ts}
\title{Estimate models}
\usage{
model(.data, ...)

\method{model}{tbl_ts}(.data, ...)
}
\arguments{
\item{.data}{A data structure suitable for the models (such as a \code{tsibble})}

\item{...}{Definitions for the models to be used}
}
\description{
Trains specified model definitions to a dataset.
}
