% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{new_decomposition}
\alias{new_decomposition}
\title{Helper to create a new decomposition function}
\usage{
new_decomposition(.class, .data, formula, ..., .env = caller_env(n = 2))
}
\arguments{
\item{defn}{R6 decomposition definition}
}
\description{
Helper to create a new decomposition function
}
