% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{new_model_definition}
\alias{new_model_definition}
\title{Define a model}
\usage{
new_model_definition(train, specials)
}
\arguments{
\item{train}{A function used to train the model to data}

\item{specials}{A list of functions used to be evaluated from the formula.
If specials is NULL, no specials are computed}
}
\description{
Define a model
}
