% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers}
\alias{outliers}
\alias{outliers.mdl_df}
\alias{outliers.mdl_ts}
\title{Identify outliers}
\usage{
outliers(object, ...)

\method{outliers}{mdl_df}(object, ...)

\method{outliers}{mdl_ts}(object, ...)
}
\arguments{
\item{object}{An object which can identify outliers.}

\item{...}{Arguments for further methods.}
}
\description{
Return a table of outlying observations using a fitted model.
}
