% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_model_lhs}
\alias{parse_model_lhs}
\title{Parse the RHS of the model formula for transformations}
\usage{
parse_model_lhs(model)
}
\arguments{
\item{model}{A model definition}
}
\description{
Parse the RHS of the model formula for transformations
}
\keyword{internal}
