% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconciliation.R
\name{reconcile}
\alias{reconcile}
\title{Forecast reconciliation}
\usage{
reconcile(data, ...)
}
\arguments{
\item{data}{A mable.}

\item{...}{Reconciliation specification}
}
\description{
Specifies the method used to reconcile forecasts across keys.
}
