% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconciliation.R
\name{reconcile}
\alias{reconcile}
\alias{reconcile.mdl_df}
\title{Forecast reconciliation}
\usage{
reconcile(.data, ...)

\method{reconcile}{mdl_df}(.data, ...)
}
\arguments{
\item{.data}{A mable.}

\item{...}{Reconciliation methods applied to model columns within \code{.data}.}
}
\description{
This function allows you to specify the method for reconciling forecasts for
hierarchical or grouped time series. \code{aggregate_key()} specifies the
relationships between the series. After an aggregated \code{tsibble} is modeled,
\code{reconcile()} defines the reconciliation algorithm, which is applied to the
individual and aggregated forecasts for each forecast horizon.
}
\details{
Forecast reconciliation ensures that the relationships between time series
(e.g. that a national forecast equals the sum of regional forecasts within
that nation) are maintained in the forecasts. Specific reconciliation methods
may have different constraints. For example, some only work with aggregations
that are summations.
}
\examples{
if (requireNamespace("fable", quietly = TRUE)) {
library(fable)
lung_deaths_agg <- as_tsibble(cbind(mdeaths, fdeaths)) \%>\%
  aggregate_key(key, value = sum(value))

lung_deaths_agg \%>\%
  model(lm = TSLM(value ~ trend() + season())) \%>\%
  reconcile(lm = min_trace(lm)) \%>\% 
  forecast()
}

}
