% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{as_tsibble}
\alias{type_sum}
\alias{obj_sum}
\alias{autoplot}
\alias{autolayer}
\alias{equation}
\alias{interpolate}
\alias{components}
\alias{augment}
\alias{glance}
\alias{tidy}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{equation}}, \code{\link[generics]{interpolate}}, \code{\link[generics]{components}}, \code{\link[generics]{augment}}, \code{\link[generics]{glance}}, \code{\link[generics]{tidy}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}, \code{\link[ggplot2]{autolayer}}}

  \item{tibble}{\code{\link[tibble]{obj_sum}}}

  \item{tsibble}{\code{\link[tsibble]{\%>\%}}, \code{\link[tsibble]{as_tsibble}}, \code{\link[tsibble]{type_sum}}}
}}

