% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{as_tsibble}
\alias{type_sum}
\alias{obj_sum}
\alias{forecast}
\alias{accuracy}
\alias{getResponse}
\alias{autoplot}
\alias{autolayer}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{forecast}{\code{\link[forecast]{forecast}}, \code{\link[forecast]{accuracy}}, \code{\link[forecast]{getResponse}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}, \code{\link[ggplot2]{autolayer}}}

  \item{tibble}{\code{\link[tibble]{obj_sum}}}

  \item{tsibble}{\code{\link[tsibble]{\%>\%}}, \code{\link[tsibble]{as_tsibble}}, \code{\link[tsibble]{type_sum}}}
}}

