% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{as_tsibble}
\alias{vars}
\alias{autoplot}
\alias{autolayer}
\alias{equation}
\alias{interpolate}
\alias{components}
\alias{augment}
\alias{glance}
\alias{tidy}
\alias{hypothesize}
\alias{generate}
\alias{refit}
\alias{hilo}
\alias{guide_level}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{distributional}{\code{\link[distributional]{guide_level}}, \code{\link[distributional]{hilo}}}

  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}, \code{\link[dplyr]{vars}}}

  \item{generics}{\code{\link[generics]{augment}}, \code{\link[generics]{components}}, \code{\link[generics]{equation}}, \code{\link[generics]{generate}}, \code{\link[generics]{glance}}, \code{\link[generics]{hypothesize}}, \code{\link[generics]{interpolate}}, \code{\link[generics]{refit}}, \code{\link[generics]{tidy}}}

  \item{ggplot2}{\code{\link[ggplot2]{autolayer}}, \code{\link[ggplot2]{autoplot}}}

  \item{tsibble}{\code{\link[tsibble:as-tsibble]{as_tsibble}}}
}}

