% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{response}
\alias{response}
\title{Extract the response variable from an object}
\usage{
response(x, ...)
}
\arguments{
\item{x}{An object containing a response variable (such as a \code{fable_model})}

\item{...}{Additional arguments for methods}
}
\value{
An expression for the response variable
}
\description{
Extract the response variable from an object
}
