% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R, R/response.R
\name{response}
\alias{response}
\title{Extract the response variable from an object}
\usage{
response(object, ...)

response(object, ...)
}
\arguments{
\item{object}{The object containing response data}

\item{...}{Additional arguments for methods}

\item{x}{An object containing a response variable}

\item{...}{Additional parameters passed on to other methods}
}
\value{
An expression for the response variable
}
\description{
Extract the response variable from an object

Extract the respone data from a model
}
