% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{response}
\alias{response}
\title{Extract the response variable from a model}
\usage{
response(object, ...)
}
\arguments{
\item{object}{The object containing response data}

\item{...}{Additional parameters passed on to other methods}
}
\description{
Returns a tsibble containing only the response variable used in the fitting
of a model.
}
