% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate responses from a model}
\usage{
simulate(object, ...)
}
\arguments{
\item{object}{A model (typically a mable)}

\item{...}{Additional optional arguments}
}
\description{
Simulate responses from a model
}
\examples{
library(fable)
library(tsibbledata)
UKLungDeaths \%>\% 
  model(lm = TSLM(mdeaths ~ fourier("year", K = 4) + fdeaths)) \%>\% 
  simulate(UKLungDeaths, times = 5)
  
}
