% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specials.R
\name{special_xreg}
\alias{special_xreg}
\title{Helper special for producing a model matrix of exogenous regressors}
\usage{
special_xreg(...)
}
\arguments{
\item{...}{Arguments for \code{fable_xreg_matrix} (see Details)}
}
\description{
Helper special for producing a model matrix of exogenous regressors
}
\details{
Currently the \code{fable_xreg_matrix} helper supports a single argument named
\code{default_intercept}. If this argument is TRUE (passed via \code{...} above), then
the intercept will be returned in the matrix if not specified (much like the
behaviour of \code{lm()}). If FALSE, then the intercept will only be included if
explicitly requested via \code{1} in the formula.
}
