% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconciliation.R
\name{top_down}
\alias{top_down}
\title{Top down forecast reconciliation}
\usage{
top_down(
  models,
  method = c("forecast_proportions", "average_proportions", "proportion_averages")
)
}
\arguments{
\item{models}{A column of models in a mable.}

\item{method}{The reconciliation method to use.}
}
\description{
\lifecycle{experimental}
}
\details{
Reconciles a hierarchy using the top down reconciliation method. The
response variable of the hierarchy must be aggregated using sums. The
forecasted time points must match for all series in the hierarchy.
}
\seealso{
\code{\link[=reconcile]{reconcile()}}, \code{\link[=aggregate_key]{aggregate_key()}}
}
